+++ src/main/java/com/ark/security/service/ProductService.java
@@ -1,5 +1,7 @@
 package com.ark.security.service;
 
+import com.ark.security.exception.NotFoundException;
+import com.ark.security.exception.NullException;
 import com.ark.security.models.Product;
 import com.ark.security.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,20 +18,34 @@ public class ProductService {
     private final ProductRepository productRepository;
 
     public Product getProductById(int id) {
-        return productRepository.findById(id).orElse(null);
+        return productRepository.findById(id).orElseThrow(()-> new NotFoundException("Không tìm thấy sản phẩm: "+ id));
     }
 
     public List<Product> getAllProducts(){
-        return productRepository.findAll();
+        List<Product> products = productRepository.findAll();
+        if(products.isEmpty()){
+            throw new NotFoundException("Không có sản phẩm nào");
+        }
+        return products;
+    }
+
+    public boolean isExistsProductByName(String name){
+        return productRepository.existsProductByName(name);
     }
 
     public void saveProduct(Product product) {
+        if(isExistsProductByName(product.getName())) {
+            throw new NotFoundException("Sản phẩm đã tồn tại");
+        }
         product.setCreatedAt(LocalDateTime.now());
         productRepository.save(product);
     }
 
     public void updateProduct(int id, Product product){
         Product oldProduct = getProductById(id);
+        if(product == null){
+            throw new NullException("Không được để trống");
+        }
         if (oldProduct != null) {
             oldProduct.setName(product.getName());
             oldProduct.setPrice(product.getPrice());
@@ -43,10 +59,16 @@ public class ProductService {
             oldProduct.setCreatedAt(LocalDateTime.now());
             oldProduct.setStatus(product.getStatus());
             productRepository.save(oldProduct);
+        }else {
+            throw new NotFoundException("Không tìm thấy sản phẩm: "+ id);
         }
     }
 
     public void deleteProduct(int id) {
+        Product product = getProductById(id);
+        if(product == null){
+            throw new NotFoundException("Không tìm thấy sản phẩm: "+ id);
+        }
         productRepository.deleteById(id);
     }
