+++ src/main/java/com/ark/security/service/SaleService.java
@@ -1,5 +1,7 @@
 package com.ark.security.service;
 
+import com.ark.security.exception.NotFoundException;
+import com.ark.security.exception.NullException;
 import com.ark.security.models.Sale;
 import com.ark.security.repository.SaleRepository;
 import lombok.RequiredArgsConstructor;
@@ -17,15 +19,22 @@ public class SaleService {
     }
 
     public List<Sale> getAllSales(){
-        return saleRepository.findAll();
+        List<Sale> sales = saleRepository.findAll();
+        if(sales.isEmpty()){
+            throw new NotFoundException("Không có khuyến mãi nào");
+        }
+        return sales;
     }
 
     public Sale getSaleById(int id){
-        return saleRepository.findById(id).orElse(null);
+        return saleRepository.findById(id).orElseThrow(()-> new NotFoundException("Không tìm thấy khuyến mãi: "+ id));
     }
 
     public void updateSale(int id, Sale sale){
         Sale oldSale = getSaleById(id);
+        if(sale == null){
+            throw new NullException("Không được để trống");
+        }
         if(oldSale !=null){
             oldSale.setName(sale.getName());
             oldSale.setDiscount(sale.getDiscount());
@@ -34,10 +43,16 @@ public class SaleService {
             oldSale.setEnd(sale.getEnd());
             oldSale.setStatus(sale.getStatus());
             saleRepository.save(oldSale);
+        }else{
+            throw new NotFoundException("Không tìm thấy khuyến mãi: "+ id);
         }
     }
 
     public void deleteSale(int id){
+        Sale sale = getSaleById(id);
+        if(sale == null){
+            throw new NotFoundException("Không tìm thấy khuyến mãi: "+ id);
+        }
         saleRepository.deleteById(id);
     }
 }
